backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai==1.3.0
pinecone==3.0.2
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.0.1
pytesseract==0.3.10
Pillow==10.1.0
google-cloud-storage==2.13.0 